+++ lib/src/simple_list/checkbox_item.dart
@@ -8,16 +8,21 @@ class CheckBoxItem<T> extends StatefulWidget {
   final OptionItem optionItem;
 
   // Função chamada quando o item do radio button é alterado
-  final ValueChanged<bool?>? onChanged;
+  final List<OptionItem>? listOptionsItem;
+
+
+  // Função chamada quando o item do radio button é alterado
+  final ValueChanged<List<OptionItem>?>? onChanged;
 
   // O estado do formulário ao qual o grupo de radio buttons pertence
-  final FormFieldState<bool> state;
+  final FormFieldState<List<OptionItem>> state;
 
   const CheckBoxItem({
     Key? key,
     required this.optionItem,
     required this.onChanged,
     required this.state,
+    required this.listOptionsItem,
   }) : super(key: key);
 
   @override
@@ -38,8 +43,9 @@ class _CheckBoxItemState extends State<CheckBoxItem> {
           value: widget.optionItem.value,
           onChanged: (value) {
             // Chama a função onChanged quando o valor é alterado
-            widget.onChanged!(value);
-            widget.state.didChange(value);
+            widget.optionItem.value = value;
+            widget.onChanged!(widget.listOptionsItem);
+            widget.state.didChange(widget.listOptionsItem);
             widget.state.validate();
           },
         ),
@@ -49,9 +55,10 @@ class _CheckBoxItemState extends State<CheckBoxItem> {
           highlightColor: Colors.transparent,
           splashColor: Colors.transparent,
           onTap: () {
+            widget.optionItem.value = !widget.optionItem.value;
             // Chama a função onChanged quando o texto é clicado
-            widget.onChanged!(widget.optionItem.value);
-            widget.state.didChange(widget.optionItem.value);
+            widget.onChanged!(widget.listOptionsItem);
+            widget.state.didChange(widget.listOptionsItem);
             widget.state.validate();
           },
           child: Text(widget.optionItem.label),