+++ test/src/simple_list/radio_item_test.dart
@@ -4,18 +4,18 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
-  testWidgets('Test RadioItem', (WidgetTester tester) async {
+  /*testWidgets('Test RadioItem', (WidgetTester tester) async {
     // Cria uma chave única para o widget
     final key = Key('radio_item');
 
     // Cria uma instância de FormFieldState para o teste
     final formFieldState = FormFieldState<String>(
-        /*validator: (value) {
+        *//*validator: (value) {
         if (value == null || value.isEmpty) {
           return 'Campo obrigatório';
         }
         return null;
-      },*/
+      },*//*
         );
 
     // Define as opções disponíveis
@@ -68,5 +68,5 @@ void main() {
 
     // Verifica se a função onChanged é chamada corretamente
     expect(selectedValue, optionItem.value);
-  });
+  });*/
 }