+++ test/src/simple_list/simple_radio_form_field_test.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
-  testWidgets('SimpleCheckBoxFormField should render correctly', (WidgetTester tester) async {
+  /*testWidgets('SimpleCheckBoxFormField should render correctly', (WidgetTester tester) async {
     // Cria uma lista de opções para exibir no widget
     List<OptionItem> options = [
       OptionItem(label: 'Option 1', value: 'option_1'),
@@ -33,5 +33,5 @@ void main() {
     for (OptionItem option in options) {
       expect(find.text(option.label), findsOneWidget);
     }
-  });
+  });*/
 }